Only pass callback to newWorker if we crash whilst starting
If the worker is already running when an unexpected disconnect occurs,
there is nothing that needs to be notified, so invoke newWorker with
no callback.

This was originally introduced to fix a different bug in which we
didn't register a successful restart if a worker didn't come up
successfully the first time.

It introduced a new bug in which it invoked a one-time-only callback
passed in by async.times during startup. When a worker died and got
restarted, the 'listening' event of that worker invoked the callback
that had already been invoked.